+++ offchain/crates/solana-fork/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- replace `spl-token` with `spl-token-interface` ([#232](https://github.com/doublezerofoundation/doublezero-offchain/pull/232))
 - use `doublezero-solana-sdk` as dependency ([#225](https://github.com/doublezerofoundation/doublezero-offchain/pull/225))
 - add god mode ([#146](https://github.com/doublezerofoundation/doublezero-offchain/pull/146))
 - add doublezero-solana-fork-cli ([#140](https://github.com/doublezerofoundation/doublezero-offchain/pull/140))